+++ apps/bare-expo/ios/Pods/Analytics/Analytics/Classes/SEGScreenPayload.m
@@ -1,0 +1,18 @@
+#import "SEGScreenPayload.h"
+
+
+@implementation SEGScreenPayload
+
+- (instancetype)initWithName:(NSString *)name
+                  properties:(NSDictionary *)properties
+                     context:(NSDictionary *)context
+                integrations:(NSDictionary *)integrations
+{
+    if (self = [super initWithContext:context integrations:integrations]) {
+        _name = [name copy];
+        _properties = [properties copy];
+    }
+    return self;
+}
+
+@end