+++ apps/bare-expo/ios/Pods/Headers/Private/Analytics/NSViewController+SEGScreen.h
@@ -1,1 +1,1 @@
-../../../Analytics/Segment/Internal/NSViewController+SEGScreen.h
+../../../Analytics/Analytics/Internal/NSViewController+SEGScreen.h