+++ apps/bare-expo/ios/Pods/Headers/Private/Analytics/SEGIdentifyPayload.h
@@ -1,1 +1,1 @@
-../../../Analytics/Segment/Classes/SEGIdentifyPayload.h
+../../../Analytics/Analytics/Classes/SEGIdentifyPayload.h