+++ apps/bare-expo/ios/Pods/Headers/Private/Analytics/SEGStoreKitTracker.h
@@ -1,1 +1,1 @@
-../../../Analytics/Segment/Internal/SEGStoreKitTracker.h
+../../../Analytics/Analytics/Internal/SEGStoreKitTracker.h