+++ apps/bare-expo/ios/Pods/Headers/Public/Analytics/SEGTrackPayload.h
@@ -1,1 +1,1 @@
-../../../Analytics/Segment/Classes/SEGTrackPayload.h
+../../../Analytics/Analytics/Classes/SEGTrackPayload.h